requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deepface==0.0.92
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
namex==0.0.8
numpy==1.26.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.0
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
pyparsing==3.1.2
PySocks==1.7.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
smmap==5.0.1
soupsieve==2.5
streamlit==1.36.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
yoloface==0.0.4
